+++ pages/login/index.tsx
@@ -1,5 +1,5 @@
 import { PageWrapper } from 'components/PageWrapper/PageWrapper'
-import LoginForm from '@/components/Login/LoginForm'
+import LoginForm from '@/components/auth/Login/LoginForm'
 import { getLayout } from '@/components/Layout/Layout'
 
 // Next запускает эту функцию, и потом рендерит страничку (в зависимости  от пропсов , которые придут). Вызывается единожды на сервере в момент билда, после страничка кешируется
